chore - extract and cleanup offset edit utils (#232475)
